// src/generated-api/src/models/HandlersCreateAffiliateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HandlersCreateAffiliateRequest
 */
export interface HandlersCreateAffiliateRequest {
    /**
     * Contact email address
     * @type {string}
     * @memberof HandlersCreateAffiliateRequest
     */
    contactEmail?: string;
    /**
     * Affiliate name
     * @type {string}
     * @memberof HandlersCreateAffiliateRequest
     */
    name: string;
    /**
     * Organization ID
     * @type {number}
     * @memberof HandlersCreateAffiliateRequest
     */
    organizationId: number;
    /**
     * Payment details in JSON format
     * swagger:strfmt json
     * @type {object}
     * @memberof HandlersCreateAffiliateRequest
     */
    paymentDetails?: object;
    /**
     * Status of the affiliate (active, pending, inactive, rejected)
     * @type {string}
     * @memberof HandlersCreateAffiliateRequest
     */
    status?: string;
}

/**
 * Check if a given object implements the HandlersCreateAffiliateRequest interface.
 */
export function instanceOfHandlersCreateAffiliateRequest(value: object): value is HandlersCreateAffiliateRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('organizationId' in value) || value['organizationId'] === undefined) return false;
    return true;
}

export function HandlersCreateAffiliateRequestFromJSON(json: any): HandlersCreateAffiliateRequest {
    return HandlersCreateAffiliateRequestFromJSONTyped(json, false);
}

export function HandlersCreateAffiliateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): HandlersCreateAffiliateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'contactEmail': json['contact_email'] == null ? undefined : json['contact_email'],
        'name': json['name'],
        'organizationId': json['organization_id'],
        'paymentDetails': json['payment_details'] == null ? undefined : json['payment_details'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function HandlersCreateAffiliateRequestToJSON(json: any): HandlersCreateAffiliateRequest {
    return HandlersCreateAffiliateRequestToJSONTyped(json, false);
}

export function HandlersCreateAffiliateRequestToJSONTyped(value?: HandlersCreateAffiliateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'contact_email': value['contactEmail'],
        'name': value['name'],
        'organization_id': value['organizationId'],
        'payment_details': value['paymentDetails'],
        'status': value['status'],
    };
}